refactor(react): move ButtonsDisabler to its own file, minor changes
