+++ packages/botonic-react/src/components/message.jsx
@@ -8,8 +8,8 @@ import { COLORS, SENDERS, WEBCHAT } from '../constants'
 import { RequestContext, WebchatContext } from '../contexts'
 import { isDev, resolveImage } from '../util/environment'
 import { ConditionalWrapper, renderComponent } from '../util/react'
-import { ButtonsDisabler } from '../util/webchat'
 import { Button } from './button'
+import { ButtonsDisabler } from './buttons-disabler'
 import { getMarkdownStyle, renderLinks, renderMarkdown } from './markdown'
 import { Reply } from './reply'
 import { MessageTimestamp, resolveMessageTimestamps } from './timestamps'
