+++ packages/botonic-react/src/node-app.jsx
@@ -5,10 +5,7 @@ import { ReactBot } from './react-bot'
 export class NodeApp {
   constructor(options) {
     this.bot = new ReactBot({
-      renderer: args => {
-        args.request.theme = options.theme
-        return this.renderNode(args)
-      },
+      renderer: args => this.renderNode(args),
       ...options,
     })
   }
